backend/requirements.txt
Flask
flask-cors
torch
numpy
Pillow
opencv-python
torchvision
scikit-image
matplotlib